lapr4/1181498-lei_isep_2019_20_sem4_2dd_1181498_1181500_1181255_1181529-3197f42c0ac1/eapli.base/base.core/src/main/java/eapli/base/product/domain/ProductCatalog.java
package eapli.base.product.domain;

import eapli.framework.validations.Preconditions;

import java.util.List;

public class ProductCatalog {

    /**
     * List of all products.
     */
    private List<Product> products;

    /**
     *
     * @param products
     */
    public ProductCatalog(List<Product> products) {
        this.products = products;
    }

    /**
     * Returns the products.
     * @return products
     */
    public List<Product> getProducts() {
        return products;
    }

    /**
     * Modifies the products.
     * @param products
     */
    public void setProducts(List<Product> products)
    {
        Preconditions.nonNull(products);
        this.products = products;
    }
}



